requirements.txt
bcrypt
cloudant
fastapi
passlib
python-slugify
uvicorn
